Normalize native credential bitmask across Windows and Unix
GWZ-Commit-ID: 01a07a9a-3f66-7786-94d1-7a1b7d841180
GWZ-Workspace-ID: ws_default
GWZ-Origin-URL-Hash: sha256:e35341f4b797aa8e57a066bf78413f40c277b3985e757118ca9aea8d390fbb9c

+++ src/git/gitbackend/transport_support.rs
@@ -142,7 +142,9 @@ pub(crate) fn remote_callbacks<'a>(
             )
         };
         if let (Some(attempt), Ok(credential)) = (&attempt, &credential) {
-            let kind = git2::CredentialType::from_bits_truncate(credential.credtype());
+            // libgit2's C enum is signed on Windows and unsigned on Unix.
+            #[allow(clippy::unnecessary_cast)]
+            let kind = git2::CredentialType::from_bits_truncate(credential.credtype() as u32);
             if !kind.is_username() {
                 let method = if identity.is_some() {
                     crate::TransportCredentialMethod::File
